// gantryweb/Frontend/Gantry/src/app/greyhound-racing/models/greyhound-racing-manual-template.model.ts
export class ManualGreyhoundRacingResponse {
    isAnyEventResulted: boolean;
    manualGreyhoundRacingRunners: ManualGreyhoundRacingRunners;
    manualGreyhoundRacingResults: ManualGreyhoundRacingResults;
}

export class ManualGreyhoundRacingRunners {
    greyhoundRacingContent: GreyhoundStaticContent;
    eventTimePlusTypeName?: string | null;
    racingContent: RacingContentData;
    greyhoundRacingEntries: Array<ManualGreyhoundRacingEntry> = [];
    isRaceOff: boolean;
    runnerCount?: string | null | undefined;
    marketEachWayString: string | null;
    gantryCommonContent: GantryCommonContent;
    isHalfScreenType: boolean;
    isFullScreenType: boolean;
    hasAnyReservedRunner: boolean;
    favPrice: number;
    isUKEvent: boolean;
    foreCast: string | null | undefined;
    triCast: string | null | undefined;
}

export class ManualGreyhoundRacingResults {
    greyhoundRacingContent: GreyhoundStaticContent;
    eventTimePlusTypeName?: string | null;
    racingContent: RacingContentData;
    time: string | null | undefined;
    eventName: string | null | undefined;
    runners: Array<ManualGreyhoundRacingResultDetails> = [];
    marketEachWayString: string | null;
    gantryCommonContent: GantryCommonContent;
    isHalfScreenType: boolean;
    isFullScreenType: boolean;
    hasAnyReservedRunner: boolean;
    isRaceOff: boolean;
    runnerCount?: string | null | undefined;
    foreCast: string | null | undefined;
    triCast: string | null | undefined;
    win: string | null | undefined;
    place: string | null | undefined;
    totes: Totes;
    eachWayResult: string | null | undefined;
    isAnyEventResulted: boolean
}

export class Totes {
    exacta: string | null | undefined;
    trifecta: string | null | undefined;
}

export class RacingContentData {
    raceNo: string | null | undefined;
    distance: string | null | undefined;
    grade: string | null | undefined;
}

export class ManualGreyhoundRacingEntry {
    trapNumber: string | null | undefined;
    greyhoundName: string | null | undefined;
    comment: string | null | undefined;
    nonRunner: boolean;
    currentPrice: string | null | undefined;
    actualPrice: number;
    jockeySilkImage: string | null | undefined;
    isReserved: boolean;
    isStartPrice: boolean;
    isVacant: boolean;
}

export class ManualGreyhoundRacingResultDetails {
    greyhoundRunnerNumber: string | null | undefined;
    greyhoundName: string | null | undefined;
    position: any | null | undefined;
    favourite: string | null | undefined;
    price: string | null | undefined;
    jockeySilkImage: string | null | undefined;
    isFavourite: boolean;
    isReserved?: boolean | null | undefined;
    trapNo: string | null | undefined;
    isStartPrice: boolean;
    currentPrice: string | null | undefined;
}